+++ desktop/tray-tauri/vite.config.js
@@ -1,0 +1,25 @@
+import { defineConfig } from "vite";
+import vue from "@vitejs/plugin-vue";
+import process from "node:process";
+
+const host = process.env.TAURI_DEV_HOST;
+
+export default defineConfig(() => ({
+  plugins: [vue()],
+  clearScreen: false,
+  server: {
+    port: 1422,
+    strictPort: true,
+    host: host || "127.0.0.1",
+    hmr: host
+      ? {
+          protocol: "ws",
+          host,
+          port: 1423,
+        }
+      : undefined,
+    watch: {
+      ignored: ["**/src-tauri/**"],
+    },
+  },
+}));